# 23-spring-data-jpa-cruddemo/src/main/java/com/luv2code/springboot/cruddemo/entity/UserAccount.java
package com.luv2code.springboot.cruddemo.entity;

import javax.persistence.Column;
import javax.persistence.Entity;
import javax.persistence.GeneratedValue;
import javax.persistence.GenerationType;
import javax.persistence.Id;
import javax.persistence.Table;
import javax.validation.constraints.Size;

@Entity
@Table(name="userAccount")
public class UserAccount {

	// define fields
	
	@Id
	@GeneratedValue(strategy=GenerationType.IDENTITY)
	@Column(name="id")
	private int id;
	
	@Size(min=2)
	@Column(name="first_name")
	private String firstName;
	
	@Column(name="last_name")
	private String lastName;
	
	@Column(name="balance")
	private double balance;
	

	
	// define constructors
	
	public UserAccount() {
		
	}

	public UserAccount(String firstName, String lastName, Double balance) {
		this.firstName = firstName;
		this.lastName = lastName;
		this.balance = balance;
		
	}

	// define getter/setter
	
	public int getId() {
		return id;
	}

	public void setId(int id) {
		this.id = id;
	}

	public String getFirstName() {
		return firstName;
	}

	public void setFirstName(String firstName) {
		this.firstName = firstName;
	}

	public String getLastName() {
		return lastName;
	}

	public void setLastName(String lastName) {
		this.lastName = lastName;
	}



	// define tostring

	public double getBalance() {
		return balance;
	}

	public void setBalance(double balance) {
		this.balance = balance;
	}

	@Override
	public String toString() {
		return "UserAccount [id=" + id + ", firstName=" + firstName
				+ ", lastName=" + lastName + ", balance=" + balance + "]";
	}


		
}