# 23-spring-data-jpa-cruddemo/target/classes/static/doc/api-docs.json
{"swagger":"2.0","info":{"description":"Awesome API Description","version":"1.0","title":"Awesome API Title","termsOfService":"urn:tos","contact":{"name":"Vincent Man","url":"https://spring.io/projects/spring-boot","email":"vincentman1027@yahoo.com.hk"},"license":{"name":"Apache 2.0","url":"http://www.apache.org/licenses/LICENSE-2.0"}},"host":"localhost:8080","basePath":"/","tags":[{"name":"web-mvc-endpoint-handler-mapping","description":"Web Mvc Endpoint Handler Mapping"},{"name":"user-account-rest-controller","description":"User Account Rest Controller"},{"name":"operation-handler","description":"Operation Handler"},{"name":"basic-error-controller","description":"Basic Error Controller"}],"consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"paths":{"/actuator":{"get":{"tags":["web-mvc-endpoint-handler-mapping"],"summary":"links","operationId":"linksUsingGET","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"$ref":"#/definitions/Map«string,Link»"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/auditevents":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/beans":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_1","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/caches":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_3","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingDELETE_1","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/actuator/caches/{cache}":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_2","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingDELETE","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}},"/actuator/conditions":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_4","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/configprops":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_5","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/env":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_7","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/env/{toMatch}":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_6","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/health":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_10","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/health/{component}":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_9","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/health/{component}/{instance}":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_8","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/heapdump":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_11","consumes":["application/xml","application/json"],"produces":["application/xml","application/octet-stream","application/json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/httptrace":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_12","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/info":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_13","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/loggers":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_15","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/loggers/{name}":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_14","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingPOST","consumes":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/mappings":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_16","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/metrics":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_18","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/metrics/{requiredMetricName}":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_17","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/scheduledtasks":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_19","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/actuator/threaddump":{"get":{"tags":["operation-handler"],"summary":"handle","operationId":"handleUsingGET_20","consumes":["application/xml","application/json"],"produces":["application/xml","application/json","application/vnd.spring-boot.actuator.v2+json"],"parameters":[{"in":"body","name":"body","description":"body","required":false,"schema":{"type":"object","additionalProperties":{"type":"string"}}}],"responses":{"200":{"description":"OK","schema":{"type":"object"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userAccount":{"put":{"tags":["user-account-rest-controller"],"summary":"updateUserAccount","operationId":"updateUserAccountUsingPUT","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"theUserAccount","description":"theUserAccount","required":true,"schema":{"$ref":"#/definitions/UserAccount"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserAccount"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userAccount/{userAccountId}/add/{amount}":{"get":{"tags":["user-account-rest-controller"],"summary":"makeTransaction","operationId":"makeTransactionUsingGET","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"parameters":[{"name":"userAccountId","in":"path","description":"userAccountId","required":true,"type":"integer","format":"int32"},{"name":"amount","in":"path","description":"amount","required":true,"type":"number","format":"double"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserAccount"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userAccounts":{"get":{"tags":["user-account-rest-controller"],"summary":"findAll","operationId":"findAllUsingGET","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/UserAccount"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"post":{"tags":["user-account-rest-controller"],"summary":"addUserAccount","operationId":"addUserAccountUsingPOST","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"parameters":[{"in":"body","name":"theUserAccount","description":"theUserAccount","required":true,"schema":{"$ref":"#/definitions/UserAccount"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserAccount"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/api/userAccounts/{userAccountId}":{"get":{"tags":["user-account-rest-controller"],"summary":"getUserAccount","operationId":"getUserAccountUsingGET","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"parameters":[{"name":"userAccountId","in":"path","description":"userAccountId","required":true,"type":"integer","format":"int32"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/UserAccount"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/error":{"get":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingGET","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"head":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingHEAD","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"post":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPOST","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"put":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPUT","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}},"delete":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingDELETE","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"options":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingOPTIONS","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}},"patch":{"tags":["basic-error-controller"],"summary":"error","operationId":"errorUsingPATCH","consumes":["application/xml","application/json"],"produces":["application/xml","application/json"],"responses":{"200":{"description":"OK","schema":{"type":"object","additionalProperties":{"type":"object"}}},"401":{"description":"Unauthorized"},"204":{"description":"No Content"},"403":{"description":"Forbidden"}}}}},"definitions":{"ModelAndView":{"type":"object","properties":{"empty":{"type":"boolean"},"model":{"type":"object"},"modelMap":{"type":"object","additionalProperties":{"type":"object"}},"reference":{"type":"boolean"},"status":{"type":"string","enum":["100 CONTINUE","101 SWITCHING_PROTOCOLS","102 PROCESSING","103 CHECKPOINT","200 OK","201 CREATED","202 ACCEPTED","203 NON_AUTHORITATIVE_INFORMATION","204 NO_CONTENT","205 RESET_CONTENT","206 PARTIAL_CONTENT","207 MULTI_STATUS","208 ALREADY_REPORTED","226 IM_USED","300 MULTIPLE_CHOICES","301 MOVED_PERMANENTLY","302 FOUND","302 MOVED_TEMPORARILY","303 SEE_OTHER","304 NOT_MODIFIED","305 USE_PROXY","307 TEMPORARY_REDIRECT","308 PERMANENT_REDIRECT","400 BAD_REQUEST","401 UNAUTHORIZED","402 PAYMENT_REQUIRED","403 FORBIDDEN","404 NOT_FOUND","405 METHOD_NOT_ALLOWED","406 NOT_ACCEPTABLE","407 PROXY_AUTHENTICATION_REQUIRED","408 REQUEST_TIMEOUT","409 CONFLICT","410 GONE","411 LENGTH_REQUIRED","412 PRECONDITION_FAILED","413 PAYLOAD_TOO_LARGE","413 REQUEST_ENTITY_TOO_LARGE","414 URI_TOO_LONG","414 REQUEST_URI_TOO_LONG","415 UNSUPPORTED_MEDIA_TYPE","416 REQUESTED_RANGE_NOT_SATISFIABLE","417 EXPECTATION_FAILED","418 I_AM_A_TEAPOT","419 INSUFFICIENT_SPACE_ON_RESOURCE","420 METHOD_FAILURE","421 DESTINATION_LOCKED","422 UNPROCESSABLE_ENTITY","423 LOCKED","424 FAILED_DEPENDENCY","426 UPGRADE_REQUIRED","428 PRECONDITION_REQUIRED","429 TOO_MANY_REQUESTS","431 REQUEST_HEADER_FIELDS_TOO_LARGE","451 UNAVAILABLE_FOR_LEGAL_REASONS","500 INTERNAL_SERVER_ERROR","501 NOT_IMPLEMENTED","502 BAD_GATEWAY","503 SERVICE_UNAVAILABLE","504 GATEWAY_TIMEOUT","505 HTTP_VERSION_NOT_SUPPORTED","506 VARIANT_ALSO_NEGOTIATES","507 INSUFFICIENT_STORAGE","508 LOOP_DETECTED","509 BANDWIDTH_LIMIT_EXCEEDED","510 NOT_EXTENDED","511 NETWORK_AUTHENTICATION_REQUIRED"]},"view":{"$ref":"#/definitions/View"},"viewName":{"type":"string"}}},"Map«string,Link»":{"type":"object","additionalProperties":{"$ref":"#/definitions/Link"}},"View":{"type":"object","properties":{"contentType":{"type":"string"}}},"UserAccount":{"type":"object","properties":{"balance":{"type":"number","format":"double"},"firstName":{"type":"string"},"id":{"type":"integer","format":"int32"},"lastName":{"type":"string"}}},"Link":{"type":"object","properties":{"href":{"type":"string"},"templated":{"type":"boolean"}}}}}